refactor: 게시물과 댓글 컴포넌트 분리
